Split unit cases, renamed files.
Test disabled logger well.
